Loki: Rename loki.bulk => loki.batch
