+++ loki/__init__.py
@@ -25,7 +25,7 @@ from loki.logging import *  # noqa
 from loki.backend import *  # noqa
 from loki.transform import * # noqa
 from loki.build import *  # noqa  # pylint: disable=redefined-builtin
-from loki.bulk import *  # noqa
+from loki.batch import *  # noqa
 from loki.lint import *  # noqa
 from loki.pragma_utils import *  # noqa
 from loki.analyse import *  # noqa
