+++ loki/batch/__init__.py
@@ -5,6 +5,6 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-from loki.bulk.scheduler import * # noqa
-from loki.bulk.item import * # noqa
-from loki.bulk.configure import * # noqa
+from loki.batch.scheduler import * # noqa
+from loki.batch.item import * # noqa
+from loki.batch.configure import * # noqa
